Add files via upload
Description
Write a program to search hotels based on the customer's budget and required duration..
Inputs:
1. Number of hotels available (Integer)
2. List of hotels with following details
a. Hotel ID (integer)
b. Minimum duration - It is represented as hours (integer). If the customer stays lesser hours than this value, then they have to pay for this whole minimum duration. If the customer stays more hours than this value, then they have to pay for the minimum duration and pay the incremental charge for the remaining duration.

c. Minimum duration cost It is represented as Indian rupees (integer). The hotel will charge this cost to the customer for the minimum duration.

d. Incremental duration It is represented as hours (integer). The additional duration offered by the hotel for the customer to stay additionally to the minimum hours. Customer have to pay for this additional duration if they stay additionally to the minimum hours. Customer can stay for as many number of additional duration in the hotel, it is just that they have to pay for each additional duration. In any additional duration, if the customer like to stay lesser than the hours mentioned for the additional duration, then they have to pay for that full additional duration to the hotel.

e. Incremental duration cost-The cost for every additional duration offered by the hotel. It is represented as Indian rupees (integer)
f. Food amount Food amount(integer)

3. Hours required by the customer-Number of hours the customer wants to stay in the hotel.
4. Budget amount- Maximum amount the customer can spend for the stay.
5. Commission amount in percentage - This is the amount the user has to pay for the hotel booking system as commission amount (integer).
6. Premium discount in percentage - This is the discount percentage offered by the hotel booking system for the premium customers (integer).
7. Premium / Non premium- This represents whether the customer is the premium customer or not. (Possible options-P-Premium customer, N-Non premium customer)
8.
Tax in percentage-Tax percentage (integer)
9. With Food/Without Food - This represents whether the customer need food or not. (Possible options - Y-With Food, N Without Food)
Output:
Type here to search
1. List of hotel ids and amount for staying separated by pipe symbol(1).
Sample Input
2
1001
24
2000
8
1000
500
1002
12
2000
4
1000
500
48
7000
20
10
N
0
N
Sample Output 1001 6000